Copilot review: home bbox filter, shared checkbox dropdowns, hover test wait
- views.py: MissionOverView uses nav_track/start_point spatial filter (not
  grid_bounds) so home-page bbox matches Missions page and API behaviour.
- project.js: Add shared setupCheckboxDropdowns(); map.js and
  map_mission_filter.js call it (with dropdownInit guard and recalc callback).
- test_ui.py: Replace fixed time.sleep(0.5) with WebDriverWait for _is_yellow()
  in test_nav_track_highlights_yellow_on_hover (poll_frequency=0.2, timeout 2s).

Declined: body.missions-page overflow-x change — keep overflow:hidden so only
the table wrapper scrolls (no body scroll).
